1. 투입 금액이 부족하면 다시 돈을 입력받는다. 2. 클래스 분리
